+++ linden/indra/mimesh/libg3d-0.0.8/plugins/import/imp_md3/imp_md3.c
@@ -116,7 +116,7 @@ gboolean md3_load_skin(G3DContext *context, G3DModel *model,
 
 	basename = g_path_get_basename(filename);
 	skinname = g_strdup_printf("%.*s_default.skin",
-		strlen(basename) - 4, basename);
+		((int) strlen(basename)) - 4, basename);
 
 	g_debug("MD3: trying to open skin file %s", skinname);
